streetflowanalyzer/traffic_analyzer.py
import os, cv2, csv, argparse, datetime
from ultralytics.models import YOLO
from tqdm import tqdm
import numpy as np
from fpdf import FPDF
from fpdf.enums import XPos, YPos
from pathlib import PosixPath, Path

from streetflowanalyzer.modules.get_valid_image_paths import get_valid_image_paths


def analyze_image_metrics(
    detections, 
    image_area, 
    model_names, 
    target_classes):
    """Анализирует результат детекции и возвращает метрики."""
    pass


def find_highlight_examples(
    all_reports, 
    top_n=3):
    """Находит наиболее показательные примеры в наборе данных."""
    pass


def generate_visualizations(
    model, 
    examples_to_visualize, 
    source_dir, 
    output_dir, 
    conf):
    """Создаёт визуальные артефакты (изображения с аннотациями) для отчёта."""
    pass


def classify_scene(
    report, 
    thresholds):
    """Классифицирует сцену на основе метрик и порогов."""
    pass


if __name__ == "__main__":
    list_dirs: list[Path] = get_valid_image_paths()
    print("Этап инициализации завершён. Среда настроена, каркас скрипта создан.")