pyproject.toml
[project]
name = "streetflowanalyzer"
version = "0.1.0"
description = "An intelligent traffic analysis system powered by computer vision. Uses YOLOv8 for vehicle detection and generates comprehensive traffic analysis reports."
authors = [
    {name = "Ollldman",email = "dodrolla@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "torch (>=2.9.1,<3.0.0)",
    "torchvision (>=0.24.1,<0.25.0)",
    "fpdf2 (>=2.8.5,<3.0.0)",
    "ultralytics (>=8.3.245,<9.0.0)",
    "tqdm (>=4.67.1,<5.0.0)"
]

[tool.poetry]
packages = [
    { include = "streetflowanalyzer" },
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
